WIP: switch encoding unsigned type normalization to Uint* types
Using Java native types for unsigned values is wasteful especially
in case of uint32 and uint64. Furthermore it leads to uncertainty
between in16/uint8, int32/uint16 and int64/uint32.

This is a partial switch, where we retain java.lang types for
signed entities, but switch the unsigned ones.

+++ yang/yang-data-impl/src/test/java/org/opendaylight/yangtools/yang/data/impl/codecs/Uint64CodecStringTest.java
@@ -11,6 +11,7 @@ import static org.junit.Assert.assertEquals;
 
 import java.math.BigInteger;
 import org.junit.Test;
+import org.opendaylight.yangtools.yang.common.Uint64;
 import org.opendaylight.yangtools.yang.data.api.codec.Uint64Codec;
 import org.opendaylight.yangtools.yang.model.util.type.BaseTypes;
 
@@ -25,7 +26,7 @@ public class Uint64CodecStringTest {
     public void testSerialize() {
         Uint64Codec<String> codec = TypeDefinitionAwareCodecTestHelper.getCodec(BaseTypes.uint64Type(),
             Uint64Codec.class);
-        assertEquals("serialize", "123456789", codec.serialize(BigInteger.valueOf(123456789)));
+        assertEquals("serialize", "123456789", codec.serialize(Uint64.valueOf(123456789)));
     }
 
     @SuppressWarnings("unchecked")
